feature: Access control

+++ src/main/java/com/finance/controller/MatchController.java
@@ -4,6 +4,7 @@ import com.finance.dto.request.FilterDTO;
 import com.finance.dto.request.MatchStatusDTO;
 import com.finance.dto.response.MatchFullDTO;
 import com.finance.model.match.Match;
+import com.finance.model.user.UserSecurityAdapter;
 import com.finance.service.match.MatchService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,10 +14,13 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
+
+import static com.finance.model.user.UserAuthority.USER;
 
 @CrossOrigin
 @RestController
@@ -46,10 +50,17 @@ public class MatchController {
     }
 
     @DeleteMapping("/match/{id}")
-    public ResponseEntity<Void> delete(@PathVariable("id") Long id) {
+    public ResponseEntity<Void> delete(@PathVariable("id") Long id,
+                                       @AuthenticationPrincipal UserSecurityAdapter user) {
         Optional<Match> possibleMatch = service.getOne(id);
 
         if (possibleMatch.isPresent()) {
+
+            //Only resource owner  can manipulate
+            if (user.id != possibleMatch.get().getOffer().getLender().getUserId()) {
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
+
             service.delete(id);
 
             return new ResponseEntity<>(HttpStatus.OK);
@@ -61,10 +72,17 @@ public class MatchController {
 
     @PatchMapping("/match/{id}")
     public ResponseEntity<MatchFullDTO> update(@PathVariable("id") Long id,
-                                                @RequestBody MatchStatusDTO matchStatusDTO) {
+                                               @RequestBody MatchStatusDTO matchStatusDTO,
+                                               @AuthenticationPrincipal UserSecurityAdapter user) {
         Optional<Match> possibleMatch = service.getOne(id);
 
         if (possibleMatch.isPresent()) {
+
+            //Only resource owner  can manipulate
+            if (user.id != possibleMatch.get().getOffer().getLender().getUserId()) {
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
+
             Match m = possibleMatch.get();
             m.setStatus(matchStatusDTO.getStatus());
             m = service.save(m);
